Extract AddressingContext from CheckoutContext
