+++ src/Sylius/Behat/Context/Ui/CheckoutContext.php
@@ -12,6 +12,7 @@
 namespace Sylius\Behat\Context\Ui;
 
 use Behat\Behat\Context\Context;
+use Sylius\Behat\Context\Ui\Shop\Checkout\AddressingContext;
 use Sylius\Behat\Page\Shop\Checkout\AddressPageInterface;
 use Sylius\Behat\Page\Shop\Checkout\CompletePageInterface;
 use Sylius\Behat\Page\Shop\Checkout\SelectPaymentPageInterface;
@@ -21,16 +22,12 @@ use Sylius\Behat\Page\SymfonyPageInterface;
 use Sylius\Behat\Page\UnexpectedPageException;
 use Sylius\Behat\Service\Resolver\CurrentPageResolverInterface;
 use Sylius\Behat\Service\SharedStorageInterface;
-use Sylius\Component\Addressing\Comparator\AddressComparatorInterface;
 use Sylius\Component\Addressing\Model\CountryInterface;
 use Sylius\Component\Core\Formatter\StringInflector;
-use Sylius\Component\Core\Model\AddressInterface;
-use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Model\ProductInterface;
 use Sylius\Component\Core\Model\ShippingMethodInterface;
 use Sylius\Component\Payment\Model\PaymentMethodInterface;
 use Sylius\Component\Promotion\Model\PromotionInterface;
-use Sylius\Component\Resource\Factory\FactoryInterface;
 use Webmozart\Assert\Assert;
 
 /**
@@ -69,19 +66,14 @@ final class CheckoutContext implements Context
     private $completePage;
 
     /**
-     * @var FactoryInterface
-     */
-    private $addressFactory;
-
-    /**
      * @var CurrentPageResolverInterface
      */
     private $currentPageResolver;
 
     /**
-     * @var AddressComparatorInterface
+     * @var AddressingContext
      */
-    private $addressComparator;
+    private $addressingContext;
 
     /**
      * @param SharedStorageInterface $sharedStorage
@@ -90,9 +82,8 @@ final class CheckoutContext implements Context
      * @param SelectPaymentPageInterface $selectPaymentPage
      * @param SelectShippingPageInterface $selectShippingPage
      * @param CompletePageInterface $completePage
-     * @param FactoryInterface $addressFactory
      * @param CurrentPageResolverInterface $currentPageResolver
-     * @param AddressComparatorInterface $addressComparator
+     * @param AddressingContext $addressingContext
      */
     public function __construct(
         SharedStorageInterface $sharedStorage,
@@ -101,9 +92,8 @@ final class CheckoutContext implements Context
         SelectPaymentPageInterface $selectPaymentPage,
         SelectShippingPageInterface $selectShippingPage,
         CompletePageInterface $completePage,
-        FactoryInterface $addressFactory,
         CurrentPageResolverInterface $currentPageResolver,
-        AddressComparatorInterface $addressComparator
+        AddressingContext $addressingContext
     ) {
         $this->sharedStorage = $sharedStorage;
         $this->homePage = $homePage;
@@ -111,9 +101,8 @@ final class CheckoutContext implements Context
         $this->selectPaymentPage = $selectPaymentPage;
         $this->selectShippingPage = $selectShippingPage;
         $this->completePage = $completePage;
-        $this->addressFactory = $addressFactory;
         $this->currentPageResolver = $currentPageResolver;
-        $this->addressComparator = $addressComparator;
+        $this->addressingContext = $addressingContext;
     }
 
     /**
@@ -121,7 +110,7 @@ final class CheckoutContext implements Context
      */
     public function iWasAtTheCheckoutSummaryStep()
     {
-        $this->iSpecifiedTheShippingAddress($this->createDefaultAddress());
+        $this->addressingContext->iSpecifiedTheShippingAddress();
         $this->iProceedOrderWithShippingMethodAndPayment('Free', 'Offline');
     }
 
@@ -132,23 +121,6 @@ final class CheckoutContext implements Context
     {
         $this->iSelectShippingMethod($shippingMethodName);
         $this->selectShippingPage->nextStep();
-    }
-
-    /**
-     * @Given I am at the checkout addressing step
-     * @When I go back to addressing step of the checkout
-     */
-    public function iAmAtTheCheckoutAddressingStep()
-    {
-        $this->addressPage->open();
-    }
-
-    /**
-     * @When I try to open checkout addressing page
-     */
-    public function iTryToOpenCheckoutAddressingPage()
-    {
-        $this->addressPage->tryToOpen();
     }
 
     /**
@@ -173,99 +145,6 @@ final class CheckoutContext implements Context
     public function iTryToOpenCheckoutCompletePage()
     {
         $this->completePage->tryToOpen();
-    }
-
-    /**
-     * @When /^I choose ("[^"]+" street) for shipping address$/
-     */
-    public function iChooseForShippingAddress(AddressInterface $address)
-    {
-        $this->addressPage->selectShippingAddressFromAddressBook($address);
-    }
-
-    /**
-     * @When /^I choose ("[^"]+" street) for billing address$/
-     */
-    public function iChooseForBillingAddress(AddressInterface $address)
-    {
-        $this->addressPage->chooseDifferentBillingAddress();
-        $this->addressPage->selectBillingAddressFromAddressBook($address);
-    }
-
-    /**
-     * @When /^I specify the shipping (address as "[^"]+", "[^"]+", "[^"]+", "[^"]+" for "[^"]+")$/
-     * @When /^I specify the shipping (address for "[^"]+" from "[^"]+", "[^"]+", "[^"]+", "[^"]+", "[^"]+")$/
-     * @When /^I (do not specify any shipping address) information$/
-     * @When /^I change the shipping (address to "[^"]+", "[^"]+", "[^"]+", "[^"]+" for "[^"]+")$/
-     */
-    public function iSpecifyTheShippingAddressAs(AddressInterface $address)
-    {
-        $key = sprintf(
-            'shipping_address_%s_%s',
-            strtolower($address->getFirstName()),
-            strtolower($address->getLastName())
-        );
-        $this->sharedStorage->set($key, $address);
-
-        $this->addressPage->specifyShippingAddress($address);
-    }
-
-    /**
-     * @When I specify shipping country province as :province
-     */
-    public function iSpecifyShippingCountryProvinceAs($province)
-    {
-        $this->addressPage->selectShippingAddressProvince($province);
-    }
-
-    /**
-     * @When I specify billing country province as :province
-     */
-    public function iSpecifyBillingCountryProvinceAs($province)
-    {
-        $this->addressPage->selectBillingAddressProvince($province);
-    }
-
-    /**
-     * @When /^I specify the billing (address as "([^"]+)", "([^"]+)", "([^"]+)", "([^"]+)" for "([^"]+)")$/
-     * @When /^I specify the billing (address for "([^"]+)" from "([^"]+)", "([^"]+)", "([^"]+)", "([^"]+)", "([^"]+)")$/
-     * @When /^I (do not specify any billing address) information$/
-     */
-    public function iSpecifyTheBillingAddressAs(AddressInterface $address)
-    {
-        $this->addressPage->chooseDifferentBillingAddress();
-
-        $key = sprintf(
-            'billing_address_%s_%s',
-            strtolower($address->getFirstName()),
-            strtolower($address->getLastName())
-        );
-        $this->sharedStorage->set($key, $address);
-
-        $this->addressPage->specifyBillingAddress($address);
-    }
-
-    /**
-     * @When /^I specified the shipping (address as "[^"]+", "[^"]+", "[^"]+", "[^"]+" for "[^"]+")$/
-     */
-    public function iSpecifiedTheShippingAddress(AddressInterface $address)
-    {
-        $this->addressPage->open();
-        $this->iSpecifyTheShippingAddressAs($address);
-
-        $key = sprintf('billing_address_%s_%s', strtolower($address->getFirstName()), strtolower($address->getLastName()));
-        $this->sharedStorage->set($key, $address);
-
-        $this->iCompleteTheAddressingStep();
-    }
-
-    /**
-     * @When I specify the email as :email
-     * @When I do not specify the email
-     */
-    public function iSpecifyTheEmail($email = null)
-    {
-        $this->addressPage->specifyEmail($email);
     }
 
     /**
@@ -308,47 +187,6 @@ final class CheckoutContext implements Context
         $lastShippingMethod = end($shippingMethods);
 
         Assert::same($shippingMethodName, $lastShippingMethod);
-    }
-
-    /**
-     * @Then I should have :countryName selected as country
-     */
-    public function iShouldHaveSelectedAsCountry($countryName)
-    {
-        Assert::eq(
-            $countryName,
-            $this->addressPage->getShippingAddressCountry(),
-            sprintf('Shipping country should be %s but is not.', $countryName)
-        );
-    }
-
-    /**
-     * @Then I should have no country selected
-     */
-    public function iShouldHaveNoCountrySelected()
-    {
-        Assert::eq(
-            'Select',
-            $this->addressPage->getShippingAddressCountry(),
-            'Shipping country should not be selected.'
-        );
-    }
-
-    /**
-     * @When I complete the addressing step
-     * @When I try to complete the addressing step
-     */
-    public function iCompleteTheAddressingStep()
-    {
-        $this->addressPage->nextStep();
-    }
-
-    /**
-     * @When I go back to store
-     */
-    public function iGoBackToStore()
-    {
-        $this->addressPage->backToStore();
     }
 
     /**
@@ -430,26 +268,11 @@ final class CheckoutContext implements Context
     }
 
     /**
-     * @When /^I proceed selecting ("[^"]+" as shipping country)$/
-     */
-    public function iProceedSelectingShippingCountry(CountryInterface $shippingCountry = null)
-    {
-        $this->addressPage->open();
-        $shippingAddress = $this->createDefaultAddress();
-        if (null !== $shippingCountry) {
-            $shippingAddress->setCountryCode($shippingCountry->getCode());
-        }
-
-        $this->addressPage->specifyShippingAddress($shippingAddress);
-        $this->addressPage->nextStep();
-    }
-
-    /**
      * @When /^I proceed selecting ("[^"]+" as shipping country) with "([^"]+)" method$/
      */
     public function iProceedSelectingShippingCountryAndShippingMethod(CountryInterface $shippingCountry = null, $shippingMethodName = null)
     {
-        $this->iProceedSelectingShippingCountry($shippingCountry);
+        $this->addressingContext->iProceedSelectingShippingCountry($shippingCountry);
 
         $this->selectShippingPage->selectShippingMethod($shippingMethodName ?: 'Free');
         $this->selectShippingPage->nextStep();
@@ -511,22 +334,6 @@ final class CheckoutContext implements Context
     }
 
     /**
-     * @When /^I proceed as guest "([^"]*)" with ("[^"]+" as shipping country)$/
-     */
-    public function iProceedLoggingAsGuestWithAsShippingCountry($email, CountryInterface $shippingCountry = null)
-    {
-        $this->addressPage->open();
-        $this->addressPage->specifyEmail($email);
-        $shippingAddress = $this->createDefaultAddress();
-        if (null !== $shippingCountry) {
-            $shippingAddress->setCountryCode($shippingCountry->getCode());
-        }
-
-        $this->addressPage->specifyShippingAddress($shippingAddress);
-        $this->addressPage->nextStep();
-    }
-
-    /**
      * @When I return to the checkout summary step
      */
     public function iReturnToTheCheckoutSummaryStep()
@@ -556,22 +363,6 @@ final class CheckoutContext implements Context
     public function iConfirmMyOrder()
     {
         $this->completePage->confirmOrder();
-    }
-
-    /**
-     * @When I specify the password as :password
-     */
-    public function iSpecifyThePasswordAs($password)
-    {
-        $this->addressPage->specifyPassword($password);
-    }
-
-    /**
-     * @When I sign in
-     */
-    public function iSignIn()
-    {
-        $this->addressPage->signIn();
     }
 
     /**
@@ -616,15 +407,6 @@ final class CheckoutContext implements Context
     }
 
     /**
-     * @Then /^I should(?:| also) be notified that the "([^"]+)" and the "([^"]+)" in (shipping|billing) details are required$/
-     */
-    public function iShouldBeNotifiedThatTheAndTheInShippingDetailsAreRequired($firstElement, $secondElement, $type)
-    {
-        $this->assertElementValidationMessage($type, $firstElement, sprintf('Please enter %s.', $firstElement));
-        $this->assertElementValidationMessage($type, $secondElement, sprintf('Please enter %s.', $secondElement));
-    }
-
-    /**
      * @Then I should be informed that my order cannot be shipped to this address
      */
     public function iShouldBeInformedThatMyOrderCannotBeShippedToThisAddress()
@@ -632,39 +414,6 @@ final class CheckoutContext implements Context
         Assert::true(
             $this->selectShippingPage->hasNoShippingMethodsMessage(),
             'Shipping page should have no shipping methods message but it does not.'
-        );
-    }
-
-    /**
-     * @Then I should be able to log in
-     */
-    public function iShouldBeAbleToLogIn()
-    {
-        Assert::true(
-            $this->addressPage->canSignIn(),
-            'I should be able to login, but I am not.'
-        );
-    }
-
-    /**
-     * @Then the login form should no longer be accessible
-     */
-    public function theLoginFormShouldNoLongerBeAccessible()
-    {
-        Assert::false(
-            $this->addressPage->canSignIn(),
-            'I should not be able to login, but I am.'
-        );
-    }
-
-    /**
-     * @Then I should be notified about bad credentials
-     */
-    public function iShouldBeNotifiedAboutBadCredentials()
-    {
-        Assert::true(
-            $this->addressPage->checkInvalidCredentialsValidation(),
-            'I should see validation error, but I do not.'
         );
     }
 
@@ -895,30 +644,6 @@ final class CheckoutContext implements Context
     }
 
     /**
-     * @Then I should be redirected to the addressing step
-     */
-    public function iShouldBeRedirectedToTheAddressingStep()
-    {
-        Assert::true(
-            $this->addressPage->isOpen(),
-            'Checkout addressing step should be opened, but it is not.'
-        );
-    }
-
-    /**
-     * @Then I should be able to go to the shipping step again
-     */
-    public function iShouldBeAbleToGoToTheShippingStepAgain()
-    {
-        $this->addressPage->nextStep();
-
-        Assert::true(
-            $this->selectShippingPage->isOpen(),
-            'Checkout shipping step should be opened, but it is not.'
-        );
-    }
-
-    /**
      * @Then I should be able to go to the complete step again
      */
     public function iShouldBeAbleToGoToTheCompleteStepAgain()
@@ -988,18 +713,6 @@ final class CheckoutContext implements Context
     }
 
     /**
-     * @Given /^I have completed addressing step with email "([^"]+)" and ("[^"]+" based shipping address)$/
-     * @When /^I complete addressing step with email "([^"]+)" and ("[^"]+" based shipping address)$/
-     */
-    public function iCompleteAddressingStepWithEmail($email, AddressInterface $address)
-    {
-        $this->addressPage->open();
-        $this->iSpecifyTheEmail($email);
-        $this->iSpecifyTheShippingAddressAs($address);
-        $this->iCompleteTheAddressingStep();
-    }
-
-    /**
      * @Then the subtotal of :item item should be :price
      */
     public function theSubtotalOfItemShouldBe($item, $price)
@@ -1059,55 +772,6 @@ final class CheckoutContext implements Context
     }
 
     /**
-     * @Then I should be on the checkout addressing step
-     */
-    public function iShouldBeOnTheCheckoutAddressingStep()
-    {
-        Assert::true(
-            $this->addressPage->isOpen(),
-            'Checkout addressing page should be opened, but it is not.'
-        );
-    }
-
-    /**
-     * @When I specify the province name manually as :provinceName for shipping address
-     */
-    public function iSpecifyTheProvinceNameManuallyAsForShippingAddress($provinceName)
-    {
-        $this->addressPage->specifyShippingAddressProvince($provinceName);
-    }
-
-    /**
-     * @When I specify the province name manually as :provinceName for billing address
-     */
-    public function iSpecifyTheProvinceNameManuallyAsForBillingAddress($provinceName)
-    {
-        $this->addressPage->specifyBillingAddressProvince($provinceName);
-    }
-
-    /**
-     * @Then I should not be able to specify province name manually for shipping address
-     */
-    public function iShouldNotBeAbleToSpecifyProvinceNameManuallyForShippingAddress()
-    {
-        Assert::false(
-            $this->addressPage->hasShippingAddressInput(),
-            'I should not be able to specify manually the province for shipping address, but I can.'
-        );
-    }
-
-    /**
-     * @Then I should not be able to specify province name manually for billing address
-     */
-    public function iShouldNotBeAbleToSpecifyProvinceNameManuallyForBillingAddress()
-    {
-        Assert::false(
-            $this->addressPage->hasBillingAddressInput(),
-            'I should not be able to specify manually the province for billing address, but I can.'
-        );
-    }
-
-    /**
      * @Then I should see :provinceName in the shipping address
      */
     public function iShouldSeeInTheShippingAddress($provinceName)
@@ -1138,22 +802,6 @@ final class CheckoutContext implements Context
             $this->selectShippingPage->hasNoAvailableShippingMethodsWarning(),
             'There should be warning about no available shipping methods, but it does not.'
         );
-    }
-
-    /**
-     * @Then /^(address "[^"]+", "[^"]+", "[^"]+", "[^"]+", "[^"]+", "[^"]+") should be filled as shipping address$/
-     */
-    public function addressShouldBeFilledAsShippingAddress(AddressInterface $address)
-    {
-        Assert::true($this->addressComparator->equal($address, $this->addressPage->getPreFilledShippingAddress()));
-    }
-
-    /**
-     * @Then /^(address "[^"]+", "[^"]+", "[^"]+", "[^"]+", "[^"]+", "[^"]+") should be filled as billing address$/
-     */
-    public function addressShouldBeFilledAsBillingAddress(AddressInterface $address)
-    {
-        Assert::true($this->addressComparator->equal($address, $this->addressPage->getPreFilledBillingAddress()));
     }
 
     /**
@@ -1230,40 +878,6 @@ final class CheckoutContext implements Context
         Assert::same(
             $discount,
             $this->completePage->getShippingPromotionDiscount($promotion->getName())
-        );
-    }
-
-    /**
-     * @return AddressInterface
-     */
-    private function createDefaultAddress()
-    {
-        /** @var AddressInterface $address */
-        $address = $this->addressFactory->createNew();
-        $address->setFirstName('John');
-        $address->setLastName('Doe');
-        $address->setCountryCode('US');
-        $address->setCity('North Bridget');
-        $address->setPostcode('93-554');
-        $address->setStreet('0635 Myron Hollow Apt. 711');
-        $address->setPhoneNumber('321123456');
-
-        return $address;
-    }
-
-    /**
-     * @param string $type
-     * @param string $element
-     * @param string $expectedMessage
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function assertElementValidationMessage($type, $element, $expectedMessage)
-    {
-        $element = sprintf('%s_%s', $type, implode('_', explode(' ', $element)));
-        Assert::true(
-            $this->addressPage->checkValidationMessageFor($element, $expectedMessage),
-            sprintf('The %s should be required.', $element)
         );
     }
