+++ src/Sylius/Behat/Resources/config/suites/ui/order/managing_orders.yml
@@ -48,6 +48,7 @@ default:
                 - sylius.behat.context.ui.admin.notification
                 - sylius.behat.context.ui.shop.currency
                 - sylius.behat.context.ui.checkout
+                - sylius.behat.context.ui.checkout.addressing
                 - sylius.behat.context.ui.channel
                 - sylius.behat.context.ui.shop.cart
             filters:
