+++ src/Sylius/Behat/Resources/config/suites/ui/taxation/applying_taxes.yml
@@ -30,6 +30,7 @@ default:
 
                 - sylius.behat.context.ui.shop.cart
                 - sylius.behat.context.ui.checkout
+                - sylius.behat.context.ui.checkout.addressing
                 - sylius.behat.context.ui.user
             filters:
                 tags: "@applying_taxes && @ui"
